+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+- Messages will only print when verbose is true
+- Removed unneeded step in argument splitting due to dependency change
+
 ## 1.0.1
 
 - Changed a dependency for splitting arguments